requirements.txt
torch
torchvision
diffusers
matplotlib
tqdm
numpy
pandas
pyyaml
Pillow
accelerate